+++ operator/pkg/resources/adminclient/adminclient_suite_test.go
@@ -1,0 +1,13 @@
+package adminclient
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestAdminClient(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "AdminClient Suite")
+}