+++ operator/CHANGELOG.md
@@ -23,6 +23,7 @@ We structure this changelog in accordance with [Keep a Changelog](https://keepac
 
 ### Added
 
+- AuthN support for operator-managed admin client when `spec.auth.usernamePassword` is configured
 - Restricted security context for init and logSidecar containers
 - Native support for arm64 hosts with multi-arch container image build targets 
 - Added `operator_state.md` documenting the cluster lifecycle states